requirement.txt
faiss-cpu
groq
langchain-groq
PyPDF
langchain_google_genai
streamlit
langchain_community
python-dotenv
pypdf